requirements.txt
numpy
pygame
PyOpenGL
pyserial